Set up simple styling for HTML Table
I tried to recycle the styles for the previous
activities.

Which is kind of a bad idea,
I don't want to deal with the
images lagging the website since there are a lot
of elements using images for their background so
I just remade the styles in a simpler manner.

The contents of the view_manager view is large,
so using a hover effect and svg for background
will be laggy. Tried it out in the last activity.

In terms of layout. I did not do much for the
layout of the website.